// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  admin: typeof admin;
  apartmentListings: typeof apartmentListings;
  colleges: typeof colleges;
  contactInfo: typeof contactInfo;
  dataIntegrity: typeof dataIntegrity;
  domain: typeof domain;
  gptMatching: typeof gptMatching;
  gptMatchingHelpers: typeof gptMatchingHelpers;
  groupMessages: typeof groupMessages;
  groupSharedListings: typeof groupSharedListings;
  groups: typeof groups;
  http: typeof http;
  lib: typeof lib;
  listingImages: typeof listingImages;
  providerProfiles: typeof providerProfiles;
  publicDbView: typeof publicDbView;
  reports: typeof reports;
  roommateMatches: typeof roommateMatches;
  roommateProfiles: typeof roommateProfiles;
  savedListings: typeof savedListings;
  seed: typeof seed;
  studentProfiles: typeof studentProfiles;
  userPhotos: typeof userPhotos;
  userSettings: typeof userSettings;
  users: typeof users;
  validation: typeof validation;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;

/**
 * A utility for referencing Convex functions in your app's internal API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = internal.myModule.myFunction;
 * ```
 */
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;